Adjusted the /kickoff endpoint to incorporate the current flow. There is also some refactoring that would rename the original main.py to distinguish between the core part and testing local version.
