+++ crewai_app/__init__.py
@@ -1,3 +1,3 @@
-from crewai_app.orchestrator import KeboolaInsightsCrew, get_status
+from crewai_app.main import KeboolaInsightsCrew, get_status
 
 __all__ = ['KeboolaInsightsCrew', 'get_status']
